Move attribute to separate project
