+++ Lombiq.HelpfulLibraries.SourceGenerators.Attributes/ConstantFromJsonAttribute.cs
@@ -1,4 +1,4 @@
-﻿namespace Lombiq.HelpfulLibraries.SourceGenerators;
+﻿namespace Lombiq.HelpfulLibraries.SourceGenerators.Attributes;
 
 [System.AttributeUsage(System.AttributeTargets.Class, AllowMultiple = true)]
 public sealed class ConstantFromJsonAttribute : System.Attribute
