+++ Lombiq.HelpfulLibraries.SourceGenerators.Attributes/Lombiq.HelpfulLibraries.SourceGenerators.Attributes.csproj
@@ -1,0 +1,10 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+      <TargetFramework>netstandard2.0</TargetFramework>
+      <IsPackable>false</IsPackable>
+      <Nullable>enable</Nullable>
+      <LangVersion>latest</LangVersion>
+    </PropertyGroup>
+
+</Project>
