+++ Lombiq.HelpfulLibraries.sln
@@ -23,6 +23,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lombiq.HelpfulLibraries.Ref
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lombiq.HelpfulLibraries.SourceGenerators", "Lombiq.HelpfulLibraries.SourceGenerators\Lombiq.HelpfulLibraries.SourceGenerators.csproj", "{EEC5970E-7AF7-4360-9C6D-6F3937AEE8E6}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lombiq.HelpfulLibraries.SourceGenerators.Attributes", "Lombiq.HelpfulLibraries.SourceGenerators.Attributes\Lombiq.HelpfulLibraries.SourceGenerators.Attributes.csproj", "{D4B0C0E2-C7B0-42C8-8EED-8DD612F655D4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -69,6 +71,10 @@ Global
 		{EEC5970E-7AF7-4360-9C6D-6F3937AEE8E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EEC5970E-7AF7-4360-9C6D-6F3937AEE8E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EEC5970E-7AF7-4360-9C6D-6F3937AEE8E6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D4B0C0E2-C7B0-42C8-8EED-8DD612F655D4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D4B0C0E2-C7B0-42C8-8EED-8DD612F655D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D4B0C0E2-C7B0-42C8-8EED-8DD612F655D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D4B0C0E2-C7B0-42C8-8EED-8DD612F655D4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
